refactoring 1.5
