+++ bot/handlers/__init__.py
@@ -1,0 +1,3 @@
+from bot.handlers import check_first, info, registration, cancellations, \
+    main_menu, schedule, session_schedule, suburbans, editor, rate, \
+    other_text, inline_handlers
